+++ clients/margin_trading/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 6.7.0 - 2026-06-09
+
+### Changed (2)
+
+- Updated `binance-common` library to version `4.0.0`
+- Updated `pyproject.toml` dependencies
+
 ## 6.6.0 - 2026-05-29
 
 ### Changed (1)